# docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VEP C2M Service API",
        "description": "VEP C2M Service API",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [
        {
            "url": "",
            "description": "clouddev"
        },
        {
            "url": "",
            "description": "sit"
        },
        {
            "url": "",
            "description": "uat"
        },
        {
            "url": "",
            "description": "prd"
        }
    ],
    "components": {
        "schemas": {
...
            "UpdateRecommendationRecordDto": {
                "type": "object",
                "properties": {}
            },
            "CreateRecommendationRecordDto": {
                "type": "object",
                "properties": {}
            },
            "UpdateRecommendedByTDCMustDto": {
                "type": "object",
                "properties": {}
            },
...
        }
    },
    "paths": {
...
        "/c2m/fairs/{fairCode}/recommendations/{id}": {
            "put": {
                "operationId": "RecommendationController_updateRecommendationRecord",
                "parameters": [
                    {
                        "name": "fairCode",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateRecommendationRecordDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/c2m/fairs/{fairCode}/recommendations": {
            "get": {
                "operationId": "RecommendationController_getRecommendationRecords",
                "parameters": [
                    {
                        "name": "fairCode",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "post": {
                "operationId": "RecommendationController_createRecommendationRecord",
                "parameters": [
                    {
                        "name": "fairCode",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateRecommendationRecordDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
        "/c2m/fairs/{fairCode}/recommendations/fiscalYear/{fiscalYear}/exhibitors/{companyCcdId}": {
            "get": {
                "operationId": "RecommendationController_findRecommendedBuyer",
                "parameters": [
                    {
                        "name": "fairCode",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fiscalYear",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "companyCcdId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/c2m/fairs/{fairCode}/recommendations/fiscalYear/{fiscalYear}/buyers": {
            "get": {
                "operationId": "RecommendationController_findRecommendedExhibitor",
                "parameters": [
                    {
                        "name": "fairCode",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fiscalYear",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/c2m/fairs/{fairCode}/recommendations/admin/fiscalYear/{fiscalYear}/buyers": {
            "post": {
                "operationId": "RecommendationController_findRecommendedExhibitorToAdmin",
                "parameters": [
                    {
                        "name": "fairCode",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fiscalYear",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
        "/c2m/fairs/{fairCode}/recommendations/es2": {
            "get": {
                "operationId": "RecommendationController_searchExhibitorAdmin",
                "parameters": [
                    {
                        "name": "fairCode",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/c2m/fairs/{fairCode}/recommendations/{fairYear}/showIfBuyerHasBMList": {
            "get": {
                "operationId": "RecommendationController_showIfBuyerHasBMList",
                "parameters": [
                    {
                        "name": "fairCode",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fairYear",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/c2m/fairs/{fairCode}/recommendations/{fairYear}/recommendedByTDC": {
            "get": {
                "operationId": "RecommendationController_findRecommendedByTDCRecord",
                "parameters": [
                    {
                        "name": "fairCode",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fairYear",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "language",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "pageNum",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "pageSize",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/c2m/fairs/{fairCode}/recommendations/updateRecommendedByTDCStatus": {
            "post": {
                "operationId": "RecommendationController_updateRecommendedByTDCRecordStatus",
                "parameters": [
                    {
                        "name": "fairCode",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateRecommendedByTDCMustDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
        "/c2m/fairs/{fairCode}/recommendations/{fairYear}/pendingRecommendation": {
            "get": {
                "operationId": "RecommendationController_findpendingRecommendationRecord",
                "parameters": [
                    {
                        "name": "fairCode",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fairYear",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "language",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/c2m/fairs/{fairCode}/recommendations/{fairYear}/totalNumOfTDCRecommend": {
            "get": {
                "operationId": "RecommendationController_getTotalNumOfTDCRecommend",
                "parameters": [
                    {
                        "name": "fairCode",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fairYear",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/c2m/fairs/{fairCode}/recommendations/createRecommendBMList": {
            "post": {
                "operationId": "RecommendationController_createRecommendBMList",
                "parameters": [
                    {
                        "name": "fairCode",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
        "/c2m/fairs/{fairCode}/recommendations/checkDuplicateExhibitors": {
            "post": {
                "operationId": "RecommendationController_checkDuplicateExhibitors",
                "parameters": [
                    {
                        "name": "fairCode",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
...
        "/c2m/fairs/{fairCode}/recommendations/fiscalYear/{fiscalYear}/ssoUid/{ssoUid}/noti": {
            "get": {
                "operationId": "RecommendationController_findRecommendedExhibitorForNoti",
                "parameters": [
                    {
                        "name": "ssoUid",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fairCode",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fiscalYear",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "c2mParticipantStatusId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
...
    }
}